src/cad2gcp/src/requirements.txt
google-cloud-storage
functions-framework